+++ engine/flink/schemed-kafka-components-utils/src/test/scala/pl/touk/nussknacker/engine/schemedkafka/helpers/KafkaAvroSpecMixin.scala
@@ -8,6 +8,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.scalatest.Assertion
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
+import pl.touk.nussknacker.engine.ModelConfig.JsonLikeValuesEnteringMode
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
@@ -107,15 +108,13 @@ trait KafkaAvroSpecMixin
   }
 
   protected def universalSinkFactory(
-                                      enableSingleParameterWithTemplateInsteadOfDynamicForm: Boolean
-                                    ): UniversalKafkaSinkFactory
+      jsonLikeValuesEnteringMode: JsonLikeValuesEnteringMode
+  ): UniversalKafkaSinkFactory = {
     val universalPayload = UniversalSchemaBasedSerdeProvider.create(schemaRegistryClientFactory, kafkaConfig)
     new UniversalKafkaSinkFactory(
       schemaRegistryClientFactory,
       universalPayload,
-      testModelConfig.copy(enableSingleParameterWithTemplateInsteadOfDynamicForm =
-        enableSingleParameterWithTemplateInsteadOfDynamicForm
-      ),
+      testModelConfig.copy(jsonLikeValuesEnteringMode = jsonLikeValuesEnteringMode),
       kafkaConfig,
       FlinkKafkaUniversalSinkImplFactory
     )