+++ engine/flink/schemed-kafka-components-utils/src/test/scala/pl/touk/nussknacker/engine/schemedkafka/schema/AllTypesAvroSchemaWithDefaultValues.scala
@@ -55,7 +55,7 @@ object AllTypesAvroSchemaWithDefaultValues extends TestSchema {
        |          "BLUE"
        |        ]
        |      },
-       |      "default": "RED"
+       |      "default": "BLUE"
        |    },
        |    {
        |      "name": "arrayField",