+++ engine/flink/schemed-kafka-components-utils/src/test/scala/pl/touk/nussknacker/engine/schemedkafka/schema/TestSchema.scala
@@ -1,13 +1,16 @@
 package pl.touk.nussknacker.engine.schemedkafka.schema
 
+import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.avro.Schema
 import org.apache.avro.generic.{GenericData, GenericRecord}
 import pl.touk.nussknacker.engine.api.validation.ValidationMode
 import pl.touk.nussknacker.engine.schemedkafka.AvroUtils
 import pl.touk.nussknacker.engine.schemedkafka.encode.ToAvroSchemaBasedEncoder
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.confluent.ConfluentUtils
 
 trait TestSchema {
-  lazy val schema: Schema = AvroUtils.parseSchema(stringSchema)
+  lazy val schema: Schema              = AvroUtils.parseSchema(stringSchema)
+  lazy val confluentSchema: AvroSchema = ConfluentUtils.convertToAvroSchema(schema)
   def stringSchema: String
 }