+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/language/json/JsonParser.scala
@@ -88,7 +88,7 @@ object JsonParser extends ExpressionParser {
 case class JsonExpressionTypingInfo(json: Json) extends ExpressionTypingInfo {
 
   override val typingResult: TypingResult = {
-    Typed.fromInstance(FromJsonSimpleDecoder.jsonToAnyWithOrderKeeping(json))
+    Typed.fromInstance(FromJsonSimpleDecoder.jsonToAny(json))
   }
 
 }