+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schema/AvroSchemaBasedParameter.scala
@@ -66,7 +66,7 @@ object AvroSchemaBasedParameter {
           )
         )
       } else {
-        val listOfValidatedParams = recordFields.sortBy(_.name()).map { recordField =>
+        val listOfValidatedParams = recordFields.map { recordField =>
           val fieldName = recordField.name()
           // Fields of nested records are flatten, e.g. { a -> { b -> _ } } => { a.b -> _ }
           val concatName = ParameterName(paramName.map(pn => s"${pn.value}.$fieldName").getOrElse(fieldName))